+++ src/requisition-validation/validation.factory.spec.js
@@ -285,6 +285,11 @@ describe('validationFactory', function() {
             expect(validationFactory.requestedQuantity(lineItem, requisitionMock)).toEqual('required');
         });
 
+        it('should return required if calculatedOrderQuantity column is present and displayed', function() {
+            calculatedOrderQuantityIsaColumn.$display = false;
+            requisitionMock.emergency = true;
+            expect(validationFactory.requestedQuantity(lineItem, requisitionMock)).toEqual('required');
+        });
     });
 
     describe('totalStockoutDays', function() {